+++ RELEASE_NOTES.md
@@ -11,6 +11,10 @@ Until then, deprecated features may be removed in the next minor version.
 ### ✨ Features
 
 * `NSTextAttachment` has a new `attachedImage` property.
+* `RichTextAttributeWriter` now supports affecting the entire text by default.
+* `RichTextColorWriter` now supports affecting the entire text by default.
+* `RichTextFontWriter` now supports affecting the entire text by default.
+* `RichTextStyleWriter` now supports affecting the entire text by default.
 
 ### 💡 Adjustments